// src/kiota/types.ts
export interface KiotaLogEntry {
  level: LogLevel;
  message: string;
}
// ...
export enum KiotaGenerationLanguage {
  // eslint-disable-next-line @typescript-eslint/naming-convention
  CSharp = 0,
  // eslint-disable-next-line @typescript-eslint/naming-convention
  Java = 1,
  // eslint-disable-next-line @typescript-eslint/naming-convention
  TypeScript = 2,
  // eslint-disable-next-line @typescript-eslint/naming-convention
  PHP = 3,
  // eslint-disable-next-line @typescript-eslint/naming-convention
  Python = 4,
  // eslint-disable-next-line @typescript-eslint/naming-convention
  Go = 5,
  // eslint-disable-next-line @typescript-eslint/naming-convention
  Swift = 6,
  // eslint-disable-next-line @typescript-eslint/naming-convention
  Ruby = 7,
  // eslint-disable-next-line @typescript-eslint/naming-convention
  CLI = 8,
  // eslint-disable-next-line @typescript-eslint/naming-convention
  Dart = 9,
};
// ...
export enum LogLevel {
  trace = 0,
  debug = 1,
  information = 2,
  warning = 3,
  error = 4,
  critical = 5,
  none = 6,
}
// ...
export enum MaturityLevel {
  experimental = 0,
  preview = 1,
  stable = 2,
}

export enum DependencyType {
  abstractions,
  serialization,
  authentication,
  http,
  bundle,
  additional,
}